Remove the Nutty `Helpers()` from the god-object. (#23)
Unmangle the crazy-ass `helpers` that get glued on the `picamera2`
object, but also maintain an internal reference to it and get called
from requests by passing the picamera2 object, so they can be...idk.
This was daffy.
