+++ tests/capture_dng_and_jpeg_helpers.py
@@ -5,6 +5,7 @@
 import time
 
 from picamera2 import Picamera2
+from picamera2.helpers import Helpers
 
 picam2 = Picamera2()
 picam2.start_preview()
@@ -19,7 +20,8 @@ time.sleep(2)
 buffers, metadata = picam2.switch_mode_and_capture_buffers(
     capture_config, ["main", "raw"]
 )
-picam2.helpers.save(
-    picam2.helpers.make_image(buffers[0], capture_config["main"]), metadata, "full.jpg"
+
+Helpers.save(
+    picam2, Helpers.make_image(buffers[0], capture_config["main"]), metadata, "full.jpg"
 )
-picam2.helpers.save(buffers[1], metadata, "full.jpeg")
+Helpers.save(picam2, buffers[1], metadata, "full.jpeg")
