+++ tests/capture_helpers.py
@@ -5,6 +5,7 @@
 import time
 
 from picamera2 import Picamera2
+from picamera2.helpers import Helpers
 
 picam2 = Picamera2()
 picam2.start_preview()
@@ -18,7 +19,7 @@ time.sleep(2)
 
 buffers, metadata = picam2.switch_mode_and_capture_buffers(capture_config, ["main"])
 
-arr = picam2.helpers.make_array(buffers[0], capture_config["main"])
-image = picam2.helpers.make_image(buffers[0], capture_config["main"])
+arr = Helpers.make_array(buffers[0], capture_config["main"])
+image = Helpers.make_image(buffers[0], capture_config["main"])
 
 picam2.close()
